更新时间触发器 修复IOCP组件中的一处错误 添加服务测试程序
git-svn-id: https://lcxl-pc/svn/lcxl-net-loader/trunk@24 cdb7d645-a706-6d45-958a-fa730e5ca134

+++ loader/driver/flt_dbg.c
@@ -57,7 +57,7 @@ filterAuditAllocMem(
     //
     // Integer overflow check
     //
-    if ((Size + sizeof(FILTERD_ALLOCATION)) < Size)
+    if ((Size + (ULONG)sizeof(FILTERD_ALLOCATION)) < Size)
     {
         DEBUGP(DL_VERY_LOUD+50,
                "filterAuditAllocMem: Integer overflow error file %d, line %d, Size %d \n",
